+++ src/components/FaqCard.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import AccordionItem from "@/components/ui/accordion/AccordionItem";
-import AccordionTrigger from "@/components/ui/accordion/AccordionTrigger";
-import AccordionContent from "@/components/ui/accordion/AccordionContent";
-import cn from "@/lib/utils";
+import AccordionItem from "@/components/accordion/AccordionItem";
+import AccordionTrigger from "@/components/accordion/AccordionTrigger";
+import AccordionContent from "@/components/accordion/AccordionContent";
+import cn from "@/lib/helpers/cn";
 import { useState } from "react";
 
 interface FaqCardProps {
@@ -24,18 +24,18 @@ export default function FaqCard({
   return (
     <AccordionItem
       value={value} //It's a require attribute for track what item is open or close
-      className={cn("relative overflow-hidden bg-white w-full ", classname)}
+      className={cn("relative w-full overflow-hidden bg-white", classname)}
     >
       <div className="relative z-10 bg-white">
         <AccordionTrigger
-          className="font-bold px-4 py-3 type-title-medium data-[state=open]:bg-yellow/20 data-[state=closed]:bg-white"
+          className="type-title-medium data-[state=open]:bg-yellow/20 px-4 py-3 font-bold data-[state=closed]:bg-white"
           onClick={() => setIsOpen(!isOpen)}
         >
-          <p className="w-full h-full">{question}</p>
+          <p className="h-full w-full">{question}</p>
         </AccordionTrigger>
       </div>
       <AccordionContent
-        className="type-body-medium z-10 px-4 pt-3 text-black text-left transition duration-300"
+        className="type-body-medium z-10 px-4 pt-3 text-left text-black transition duration-300"
         isOpen={isOpen}
       >
         <p>{answer}</p>
