+++ src/components/Icon.tsx
@@ -1,4 +1,4 @@
-import cn from "@/lib/utils";
+import cn from "@/lib/helpers/cn";
 
 type IconProps = {
   name: string;
