+++ src/components/accordion/AccordionItem.tsx
@@ -1,5 +1,5 @@
 import * as AccordionPrimitive from "@radix-ui/react-accordion";
-import cn from "@/lib/utils";
+import cn from "@/lib/helpers/cn";
 
 export default function AccordionItem({
   className,
