+++ src/components/accordion/AccordionTrigger.tsx
@@ -1,5 +1,5 @@
 import * as AccordionPrimitive from "@radix-ui/react-accordion";
-import cn from "@/lib/utils";
+import cn from "@/lib/helpers/cn";
 import Icon from "@/components/Icon";
 
 export default function AccordionTrigger({
@@ -12,7 +12,7 @@ export default function AccordionTrigger({
       <AccordionPrimitive.Trigger
         data-slot="accordion-trigger"
         className={cn(
-          "flex flex-1 items-start justify-between gap-4 py-4 text-left text-sm font-medium transition-all outline-none disabled:pointer-events-none disabled:opacity-50 [&[data-state=open]>i]:rotate-180 ",
+          "flex flex-1 items-start justify-between gap-4 py-4 text-left text-sm font-medium transition-all outline-none disabled:pointer-events-none disabled:opacity-50 [&[data-state=open]>i]:rotate-180",
           className
         )}
         {...props}
@@ -21,7 +21,7 @@ export default function AccordionTrigger({
 
         <Icon
           name="keyboard_arrow_down"
-          className="text-muted-foreground pointer-events-none text-red size-4 shrink-0 translate-y-0.5 transition-all duration-200"
+          className="text-muted-foreground text-red pointer-events-none size-4 shrink-0 translate-y-0.5 transition-all duration-200"
         />
       </AccordionPrimitive.Trigger>
     </AccordionPrimitive.Header>
